Create raw_transcription.txt
There was a presentation given by representatives of DEFCON Incorporated on 29 May 2025, on the DEFCON Discord server.

Representatives from the company that runs the DEFCON conference were the slide deck presenters. They entertained some questions at the end from the people that were able to attend the Discord event. The event was announced the previous day. The Town Hall was held to answer concerns from an announcement made in early May 2025 on some social media platforms, about intended changes to DEFCON Groups.

An mp4 file of the event was recorded. This is a raw transcription of that file. This transcription was made with the transcription function in Microsoft Word by @Abraxas3d